+++ packages/core/types/types/index.d.ts
@@ -1,4 +1,4 @@
 export { RegisterInteraction, GetInteraction, Interaction, Interactions } from "./interaction";
 export { Route, RouteDescriptor, ParamParser, ParamParsers, ResponseBuilder, EveryMatchFn, InitialMatchFn, OnFns } from "./route";
-export { Response, RawParams, Params, MatchResponseProperties, SettableResponseProperties } from "./response";
+export { Response, RawParams, Params, Resolved, MatchResponseProperties, SettableResponseProperties } from "./response";
 export { CuriRouter, RouterOptions, Observer, Emitted, RemoveObserver, SideEffect, Navigation } from "./curi";